+++ exporters/secret/exporter.go
@@ -5,6 +5,7 @@ import (
 	"crypto/x509"
 	"encoding/pem"
 	"fmt"
+	"strings"
 
 	"github.com/giantswarm/k8sclient/v5/pkg/k8srestconfig"
 	"github.com/giantswarm/microerror"
@@ -12,6 +13,9 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 )
@@ -25,10 +29,11 @@ type Config struct {
 }
 
 type Exporter struct {
-	cert      *prometheus.Desc
-	ctx       context.Context
-	k8sClient *kubernetes.Clientset
-	logger    micrologger.Logger
+	cert          *prometheus.Desc
+	ctx           context.Context
+	k8sClient     *kubernetes.Clientset
+	logger        micrologger.Logger
+	dynamicClient dynamic.Interface
 
 	namespaces []string
 }
@@ -81,6 +86,11 @@ func (e *Exporter) calculateExpiry(ch chan<- prometheus.Metric, secret v1.Secret
 	secretNamespace := secret.Namespace
 	certKeys := [...]string{"ca.crt", "tls.crt"}
 
+	certCRName, err := e.findCertCRName(secretName, secretNamespace)
+	if err != nil {
+		e.logger.Log("error", microerror.Mask(err))
+	}
+
 	for _, certKey := range certKeys {
 		certBytes, ok := secret.Data[certKey]
 		if !ok {
@@ -101,12 +111,67 @@ func (e *Exporter) calculateExpiry(ch chan<- prometheus.Metric, secret v1.Secret
 
 		for _, cert := range certs {
 			timestamp := float64(cert.NotAfter.Unix())
-			ch <- prometheus.MustNewConstMetric(e.cert, prometheus.GaugeValue, timestamp, secretName, secretNamespace, certKey, certType)
+			issuer := cert.Issuer.String()
+			ch <- prometheus.MustNewConstMetric(e.cert, prometheus.GaugeValue, timestamp, secretName, secretNamespace, certKey, issuer, certCRName)
 		}
 	}
 	e.logger.Log("info", fmt.Sprintf("added secret %s/%s to the metrics", secretNamespace, secretName))
 
 	return nil
+}
+
+func (e *Exporter) certificateGroupVersionResources() ([]schema.GroupVersionResource, error) {
+	groupVersions := []schema.GroupVersionResource{}
+
+	_, resources, err := e.k8sClient.ServerGroupsAndResources()
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	for _, rl := range resources {
+		for _, r := range rl.APIResources {
+			if r.Name == "certificates" && strings.HasPrefix(rl.GroupVersion, "cert-manager.io/") {
+				s := schema.FromAPIVersionAndKind(rl.GroupVersion, r.Name)
+				groupVersions = append(groupVersions, schema.GroupVersionResource{
+					Group:    s.Group,
+					Version:  s.Version,
+					Resource: r.Name,
+				})
+			}
+		}
+	}
+
+	return groupVersions, nil
+}
+
+func (e *Exporter) findCertCRName(secretName, secretNamespace string) (string, error) {
+	certificateGroupVersionResources, err := e.certificateGroupVersionResources()
+	if err != nil {
+		e.logger.Log("error", microerror.Mask(err))
+		return "", err
+	}
+	if len(certificateGroupVersionResources) == 0 {
+		e.logger.Log("info", "cert-manager Certificate not available, skipping secret collection")
+		return "", nil
+	}
+
+	for _, gvr := range certificateGroupVersionResources {
+		certs, err := e.dynamicClient.Resource(gvr).Namespace(secretNamespace).List(e.ctx, metav1.ListOptions{})
+		if err != nil {
+			e.logger.Log("error", microerror.Mask(err))
+		}
+		for _, cert := range certs.Items {
+			name, found, err := unstructured.NestedString(cert.UnstructuredContent(), "spec", "secretName")
+			if err != nil {
+				e.logger.Log("error", microerror.Mask(err))
+			}
+			if found && name == secretName {
+				return cert.GetName(), nil
+			}
+		}
+	}
+
+	return "", nil
 }
 
 func (e *Exporter) Describe(ch chan<- *prometheus.Desc) {
@@ -140,6 +205,11 @@ func New(config Config) (*Exporter, error) {
 		return nil, err
 	}
 
+	dynClient, err := dynamic.NewForConfig(restConfig)
+	if err != nil {
+		return nil, err
+	}
+
 	ctx := context.Background()
 
 	logger.Log("info", "creating new exporter")
@@ -153,12 +223,14 @@ func New(config Config) (*Exporter, error) {
 				"namespace",
 				"secretkey",
 				"issuer",
+				"crname",
 			},
 			nil,
 		),
-		ctx:        ctx,
-		k8sClient:  k8sClient,
-		logger:     logger,
-		namespaces: config.Namespaces,
+		ctx:           ctx,
+		k8sClient:     k8sClient,
+		dynamicClient: dynClient,
+		logger:        logger,
+		namespaces:    config.Namespaces,
 	}, nil
 }